+++ checker-fail-cwl.json
@@ -1,0 +1,7 @@
+{
+  "input_file": {
+        "class": "File",
+        "path": "md5sum.input"
+    },
+    "expected_md5": "not a matching md5sum"
+}
